chef-pedant 1.0.1
This allows us to pick up the 'knife' executable from the PATH, which
ensures that we can test a client that matches the given version of
the server.

Added some environment tweaks to the shelled-out commands to ensure
that Bundler doesn't poison those commands (knife would fail because
chef couldn't be found in the bundle)

One of our knife tests was slightly tweaked to reflect slightly
different behavior on newer knife versions.