The webui is now disabled by default in Chef Server
The webui is no longer enabled by default. This change reflects the
start of the official deprecation of the webui as a part of Chef
Server. It is also a better default since the webui initializes with a
known default password. Users who do not immediately reset the admin
password could be vulnerable to a trivial attack if the webui is
accessible.