add configurable bookshelf url attribute
This new attribute will default to the value of the Nginx url which is
built from the configured `api_fqdn` and Nginx ssl port.

Values set in the `/etc/chef-server/chef-server.rb` file always take
precedence so it is still possible to change the bookshelf vip to
something like "https://s3.amazonaws.com" if S3 is being used as the
backend cookbook store.

This fixes the following issues:

* [CHEF-3853] checksum URLs generated by POST /sandboxes do not respect
  configured load balancer port