fix windows version pinning in chef-dk
the lack of chef-windows version pin meant that chefdk on windows
was appbundler'ing chef-12 command line tools.  this may be the
root cause of all the @#$@#$ annoying omnitruck failures from people
trying to use knife 12.x and failing to bootstrap.

also bumped the chef version to 11.16.0 consistently and dropped the
old nokogiri dep