Do not wait for service to start
If the service fails to start (for example because the config is
missing) the installer will fail. This creates an ordering dependency,
as well as an opaque error.

Specifically we get error 1603 or error 1920:
Error 1920. Service 'Pushy Client Service' (pushy-client) failed to start.
Verify that you have sufficient privileges to start system services

By skipping the wait we can continue installation even if the service
fails to start. However this requires on the cookbook to restart the
service after the correct config file is in place; otherwise the
service will be offline.