Let omnibus pick number of workers, except in a Solaris zone.
When running in a Solaris zone with Ohai 8, the total number of
CPUs reported is 0. Instead of overriding the number of workers
globally, restrict it to just Solaris and a total CPU count of 0.

Also enables UTF8 encoding on all platforms (because some don't
by default and this breaks many things)